Fixed wind speed risk panel

+++ apps/web/app/_utils/transform.tsx
@@ -11,8 +11,8 @@ export function prepareWindChartData(f: Array<any>) : Array<any> {
 export function prepareProbabilitiesData(f: Array<any>) : Array<any> {
   return f.map(d => ({
     time: d.valid_time,
-    p_450: parseInt(d.p_v_ge_450),
-    p_500: parseInt(d.p_v_ge_500),
-    p_600: parseInt(d.p_v_ge_600),
+    p_450: parseFloat(d.p_v_ge_450),
+    p_500: parseFloat(d.p_v_ge_500),
+    p_600: parseFloat(d.p_v_ge_600),
   }));
 }
